refactor(api): extract orgs from client.ts
